03/12/2024
.

+++ app/routes.py
@@ -7,7 +7,45 @@ from flask import current_app as app
 from .email_otp import send_otp
 from functools import wraps
 
-# Decorator to check if the user is logged in and has the required role
+roles_permissions = {
+    "admin": [
+        "view_warehouse", "add_warehouse", "edit_warehouse", "delete_warehouse",
+        "view_products", "add_product", "edit_product", "delete_product",
+        "view_raw_material", "add_raw_material", "edit_raw_material", "delete_raw_material",
+        "view_users", "add_user", "edit_user", "delete_user",
+        "view_contractors", "edit_contractor", "delete_contractor",
+        "view_production_lines", "add_production_line", "edit_production_line", "delete_production_line",
+        "view_supervisors", "edit_supervisor", "delete_supervisor",
+        "view_product_raw_materials", "add_product_raw_material", "edit_product_raw_material", "delete_product_raw_material"
+    ],
+    "manager": [
+        "view_warehouse", "add_warehouse", "edit_warehouse", "delete_warehouse",
+        "view_products", "add_product", "edit_product", "delete_product",
+        "view_raw_material", "add_raw_material", "edit_raw_material", "delete_raw_material",
+        "view_contractors", "edit_contractor", "delete_contractor",
+        "view_production_lines", "add_production_line", "edit_production_line", "delete_production_line",
+        "view_supervisors", "edit_supervisor", "delete_supervisor",
+        "view_product_raw_materials", "add_product_raw_material", "edit_product_raw_material", "delete_product_raw_material"
+    ],
+    "staff": [
+        "view_warehouse", "view_products", "view_raw_material", "view_contractors",
+        "view_production_lines", "view_supervisors", "view_product_raw_materials"
+    ],
+    "supervisor": [
+        "view_products", "view_raw_material", "view_product_raw_materials"
+    ],
+    "contractor": [
+        "view_products", "view_raw_material"
+    ]
+}
+
+
+@app.context_processor
+def inject_permissions():
+    if current_user.is_authenticated:
+        return {"username": current_user.username, "user_permissions": roles_permissions.get(current_user.role, [])}
+    return {"username": "Guest", "permissions": []}
+
 def role_required(*roles):
     def decorator(func):
         @wraps(func)
@@ -37,6 +75,14 @@ def signup():
             user.set_password(form.password.data)
             db.session.add(user)
             db.session.commit()
+            if form.role.data == 'contractor':
+                contractor = Contractor(user_id=user.id)
+                db.session.add(contractor)
+                db.session.commit()
+            elif form.role.data == 'supervisor':    
+                supervisor = Supervisor(supervisor_name=user.username)
+                db.session.add(supervisor)
+                db.session.commit()
         else:
             flash('Invalid input', 'danger')
             return render_template('signup.html', form=form)
@@ -90,11 +136,14 @@ def dashboard():
         "total_users": 120,
         "active_users": 85,
         "new_users": 10,
-        "roles": {"Admin": 5, "Editor": 15, "Viewer": 100},
         "recent_activity": ["Added user 'Alice'", "Updated role for 'Bob'"],
     }
     return render_template('dashboard.html', username=current_user.username,data=data)
 
+
+#---------------------------------------Raw Material----------------------------------------------
+
+# View Raw Material
 @app.route('/raw_material')
 @login_required
 def view_raw_material():
@@ -102,7 +151,6 @@ def view_raw_material():
     print(entries)
     return render_template('raw_material.html', raw_materials=entries)
 
-#---------------------------------------Raw Material----------------------------------------------
 # Add Raw Material
 @app.route('/add_raw_material', methods=['GET', 'POST'])
 @login_required
@@ -150,6 +198,10 @@ def edit_raw_material(material_id):
 @role_required('admin','manager')
 def delete_raw_material(material_id):
     material = RawMaterial.query.get_or_404(material_id)
+    dependencies = ProductRawMaterial.query.filter_by(raw_material_id=material_id).all()
+    if dependencies:
+        flash('Cannot delete raw material. It is currently being used in one or more products.', 'warning')
+        return redirect(url_for('view_raw_material'))
     db.session.delete(material)
     db.session.commit()
     flash("Raw material deleted successfully!", "success")
@@ -205,6 +257,10 @@ def edit_product(product_id):
 @role_required('admin','manager')
 def delete_product(product_id):
     product = Product.query.get_or_404(product_id)
+    dependencies = ProductRawMaterial.query.filter_by(product_id=product_id).all()
+    if dependencies:
+        flash('Cannot delete product. It is currently linked to raw materials.', 'warning')
+        return redirect(url_for('view_products'))
     db.session.delete(product)
     db.session.commit()
     flash("Product deleted successfully!", "success")
@@ -276,11 +332,23 @@ def view_users():
 @role_required('admin')
 def add_user():
     form = SignupForm()
+    form._fields.pop('verification_code')
+    form.submit.label.text = "Confirm"
     if form.validate_on_submit():
         user = User(username=form.username.data, email=form.email.data, role=form.role.data)
         user.set_password(form.password.data)
         db.session.add(user)
         db.session.commit()
+        if form.role.data == 'contractor':
+            contractor = Contractor(user_id=user.id)
+            db.session.add(contractor)
+            db.session.commit()
+        
+        if form.role.data == 'supervisor':
+            supervisor = Supervisor(supervisor_name=user.username)
+            db.session.add(supervisor)
+            db.session.commit()
+
         flash("User added successfully!", "success")
         return redirect(url_for('view_users'))
     return render_template('form.html', form=form)
@@ -292,6 +360,10 @@ def add_user():
 def edit_user(user_id):
     user = User.query.get_or_404(user_id)
     form = SignupForm(obj=user)
+    # form._fields.pop('password')
+    # form._fields.pop('confirm_password')
+    form._fields.pop('verification_code')
+    form.submit.label.text = "Confirm"
     if form.validate_on_submit():
         user.username = form.username.data
         user.email = form.email.data
@@ -314,17 +386,206 @@ def delete_user(user_id):
     flash("User deleted successfully!", "success")
     return redirect(url_for('view_users'))
 
+
+#---------------------------------------Contractor----------------------------------------------
+@app.route('/contractors')
+@login_required
+@role_required('admin','manager')
+def view_contractors():
+    entries = Contractor.query.all()
+    print(entries)
+    return render_template('contractor.html', contractors=entries)
+
+# Edit Contractor
+@app.route('/edit_contractor/<int:contractor_id>', methods=['GET', 'POST'])
+@login_required
+@role_required('admin','manager')
+def edit_contractor(contractor_id):
+    contractor = Contractor.query.get_or_404(contractor_id)
+    form = ContractorForm(obj=contractor)
+    form.submit.label.text = "Confirm"
+    if form.validate_on_submit():
+        contractor.user_id = contractor_id
+        contractor.contract_start_date = form.contract_start_date.data
+        contractor.contract_end_date = form.contract_end_date.data
+        db.session.commit()
+        flash("Contractor updated successfully!", "success")
+        return redirect(url_for('view_contractors'))
+    return render_template('form.html', form=form, contractor=contractor)
+
+# Delete Contractor
+@app.route('/delete_contractor/<int:contractor_id>', methods=['POST'])
+@login_required
+@role_required('admin','manager')
+def delete_contractor(contractor_id):
+    contractor = Contractor.query.get_or_404(contractor_id)
+    db.session.delete(contractor)
+    db.session.commit()
+    flash("Contractor deleted successfully!", "success")
+    return redirect(url_for('view_contractors'))
+
+
+#---------------------------------------Production Line----------------------------------------------
+
+# View Production Line
+@app.route('/production_lines')
+@login_required
+def view_production_lines():
+    entries = ProductionLine.query.all()
+    print(entries)
+    return render_template('production_line.html', production_lines=entries)
+
+# Add Production Line
+@app.route('/add_production_line', methods=['GET', 'POST'])
+@login_required
+@role_required('admin','manager')
+def add_production_line():
+    form = ProductionLineForm()
+    if form.validate_on_submit():
+        new_production_line = ProductionLine(
+            line_name=form.line_name.data
+        )
+        db.session.add(new_production_line)
+        db.session.commit()
+        flash("Production line added successfully!", "success")
+        return redirect(url_for('view_production_lines'))
+    return render_template('form.html', form=form)
+
+# Edit Production Line
+@app.route('/edit_production_line/<int:production_line_id>', methods=['GET', 'POST'])
+@login_required
+@role_required('admin','manager')
+def edit_production_line(production_line_id):
+    production_line = ProductionLine.query.get_or_404(production_line_id)
+    form = ProductionLineForm(obj=production_line)
+    form.submit.label.text = "Confirm"
+    if form.validate_on_submit():
+        production_line.line_name = form.line_name.data
+        db.session.commit()
+        flash("Production line updated successfully!", "success")
+        return redirect(url_for('view_production_lines'))
+    return render_template('form.html', form=form, production_line=production_line)
+
+# Delete Production Line
+@app.route('/delete_production_line/<int:production_line_id>', methods=['POST'])
+@login_required
+@role_required('admin','manager')
+def delete_production_line(production_line_id):
+    production_line = ProductionLine.query.get_or_404(production_line_id)
+    db.session.delete(production_line)
+    db.session.commit()
+    flash("Production line deleted successfully!", "success")
+    return redirect(url_for('view_production_lines'))
+
+#---------------------------------------Supervisor----------------------------------------------
+@app.route('/supervisors')
+@login_required
+def view_supervisors():
+    entries = Supervisor.query.all()
+    print(entries)
+    return render_template('supervisor.html', supervisors=entries)
+
+# Edit Supervisor
+@app.route('/edit_supervisor/<int:supervisor_id>', methods=['GET', 'POST'])
+@login_required
+@role_required('admin','manager')
+def edit_supervisor(supervisor_id):
+    supervisor = Supervisor.query.get_or_404(supervisor_id)
+    form = SupervisorForm(obj=supervisor)
+    form.submit.label.text = "Confirm"
+    if form.validate_on_submit():
+        supervisor.supervisor_name = form.supervisor_name.data
+        supervisor.contact_info = form.contact_info.data
+        supervisor.contractor_id = form.contractor_id.data
+        db.session.commit()
+        flash("Supervisor updated successfully!", "success")
+        return redirect(url_for('view_supervisors'))
+    return render_template('form.html', form=form, supervisor=supervisor)
+
+# Delete Supervisor
+@app.route('/delete_supervisor/<int:supervisor_id>', methods=['POST'])
+@login_required
+@role_required('admin','manager')
+def delete_supervisor(supervisor_id):
+    supervisor = Supervisor.query.get_or_404(supervisor_id)
+    db.session.delete(supervisor)
+    db.session.commit()
+    flash("Supervisor deleted successfully!", "success")
+    return redirect(url_for('view_supervisors'))
+
+#---------------------------------------Product Raw Material----------------------------------------------
+@app.route('/product_raw_materials')
+@login_required
+def view_product_raw_materials():
+    entries = ProductRawMaterial.query.all()
+    print(entries)
+    return render_template('product_raw_material.html', product_raw_materials=entries)
+
+# Add Product Raw Material
+@app.route('/add_product_raw_material', methods=['GET', 'POST'])
+@login_required
+@role_required('admin','manager')
+def add_product_raw_material():
+    form = ProductRawMaterialForm()
+    if form.validate_on_submit():
+        new_product_raw_material = ProductRawMaterial(
+            product_id=form.product_id.data,
+            raw_material_id=form.raw_material_id.data,
+            quantity_required=form.quantity_required.data
+        )
+        db.session.add(new_product_raw_material)
+        db.session.commit()
+        flash("Product raw material added successfully!", "success")
+        return redirect(url_for('view_product_raw_materials'))
+    return render_template('form.html', form=form)
+
+# Edit Product Raw Material
+@app.route('/edit_product_raw_material/<int:product_id>/<int:raw_material_id>', methods=['GET', 'POST'])
+@login_required
+@role_required('admin','manager')
+def edit_product_raw_material(product_id, raw_material_id):
+    product_raw_material = ProductRawMaterial.query.get_or_404((product_id, raw_material_id))
+    form = ProductRawMaterialForm(obj=product_raw_material)
+    form.submit.label.text = "Confirm"
+    if form.validate_on_submit():
+        product_raw_material.product_id = form.product_id.data
+        product_raw_material.raw_material_id = form.raw_material_id.data
+        product_raw_material.quantity_required = form.quantity_required.data
+        db.session.commit()
+        flash("Product raw material updated successfully!", "success")
+        return redirect(url_for('view_product_raw_materials'))
+    return render_template('form.html', form=form, product_raw_material=product_raw_material)
+
+# Delete Product Raw Material
+@app.route('/delete_product_raw_material/<int:product_id>/<int:raw_material_id>', methods=['POST'])
+@login_required
+@role_required('admin','manager')
+def delete_product_raw_material(product_id, raw_material_id):
+    product_raw_material = ProductRawMaterial.query.get_or_404((product_id, raw_material_id))
+    db.session.delete(product_raw_material)
+    db.session.commit()
+    flash("Product raw material deleted successfully!", "success")
+    return redirect(url_for('view_product_raw_materials'))
+
+#---------------------------------------Profile----------------------------------------------
+# @app.route('/profile')
+# @login_required
+# def profile():
+#     return render_template('profile.html', user=current_user)
+
+
+
 #---------------------------------------Error Handling----------------------------------------------
-@app.errorhandler(403)
-def forbidden(e):
-    return render_template('403.html'), 403
+# @app.errorhandler(403)
+# def forbidden(e):
+#     return render_template('403.html'), 403
 
-@app.errorhandler(404)
-def page_not_found(e):
-    return render_template('404.html'), 404
+# @app.errorhandler(404)
+# def page_not_found(e):
+#     return render_template('404.html'), 404
 
-@app.errorhandler(500)
-def server_error(e):
-    return render_template('500.html'), 500
+# @app.errorhandler(500)
+# def server_error(e):
+#     return render_template('500.html'), 500
 
 #---------------------------------------Miscellaneous----------------------------------------------
